move page renders to components
